requirements.txt
streamlit
numpy
opencv-python-headless
pillow
protobuf
ultralytics
torch
torchvision
torchaudio
openai